+++ libsplinter/src/oauth/rest_api/mod.rs
@@ -19,7 +19,7 @@ mod actix;
 mod resources;
 
 use crate::biome::OAuthUserSessionStore;
-use crate::rest_api::{Resource, RestResourceProvider};
+use crate::rest_api::actix_web_1::{Resource, RestResourceProvider};
 
 use super::OAuthClient;
